refactored, but not tested
