--- lib/mt/ArgObjComparison.js
@@ -1,11 +1,0 @@
-.factory                    .factory
-.prototype.getReqParams     .getReqParams
-.prototype.getRspParams
-.prototype.getMtEvtPacket
-
-
-.getType
-.getCmdNumber 
-.makeArgObj
-.transToArgObj
-.getCmdBuf
